+++ src/Kripke/Sweep_Solver.cpp
@@ -108,7 +108,7 @@ int SweepSolver (Grid_Data *grid_data, bool block_jacobi)
       if(true){
         // Create a list of all groups
         std::vector<int> sdom_list(grid_data->subdomains.size());
-        for(int i = 0;i < grid_data->subdomains.size();++ i){
+        for(size_t i = 0;i < grid_data->subdomains.size();++ i){
           sdom_list[i] = i;
         }
 
@@ -163,7 +163,7 @@ void SweepSubdomains (std::vector<int> subdomain_list, Grid_Data *grid_data, boo
   }
 
   // Add all subdomains in our list
-  for(int i = 0;i < subdomain_list.size();++ i){
+  for(size_t i = 0;i < subdomain_list.size();++ i){
     int sdom_id = subdomain_list[i];
     comm->addSubdomain(sdom_id, grid_data->subdomains[sdom_id]);
   }
